+++ src/app/search-result/search-result.component.html
@@ -1,6 +1,11 @@
 <div class="search-result">
-  <h2>検索結果一覧 ヒット件数: {{searchResult?.nbHits}}</h2>
-  <ul>
-    <li *ngFor="let hit of searchResult?.hits">{{hit.title}}</li>
-  </ul>
+  <div class="search-result__header">
+    <h2>キーワード： {{ searchQuery }}</h2>
+    <h2 class="search-result__hits">結果： {{searchResult?.nbHits}} 件見つかりました</h2>
+  </div>
+  <div class="search-result__body">
+    <ng-container *ngFor="let article of articles$ | async">
+      <app-article *ngIf="article" [article]="article"></app-article>
+    </ng-container>
+  </div>
 </div>
