+++ src/app/search-result/search-result.component.scss
@@ -1,5 +1,14 @@
 .search-result {
   max-width: 910px;
   margin: 0 auto;
-  text-align: center;
+  &__header {
+    text-align: center;
+  }
+  &__hits {
+    font-size: 20px;
+  }
+  &__body {
+    max-width: 610px;
+    margin: 0 auto;
+  }
 }
